(worldmap_mapmu) Decompile MapMu proc

+++ src/worldmap_gmapunit.c
@@ -446,7 +446,7 @@ u16 sub_80BB188(struct GMapUnitContainerProc * container, int index)
 }
 
 //! FE8U = 0x080BB194
-u16 sub_80BB194(struct GMapUnitContainerProc * container, int index)
+int sub_80BB194(struct GMapUnitContainerProc * container, int index)
 {
     return container->pMapUnitProcs[index]->unk_3a;
 }
